Added 2 new tests with system_calls
Added and fixed system function SYS_DISK_SPACE.
And now in loop we are search all files in a disk and show in console.
We are receive the disk size from "devices/block.c" in function "block_register".
The value how we are receive, we are save in a global variable how situated in syscall.h and syscall.c